1.Fixed bug that allow generation all items with itemset field value '0' at ".additemset 0" 2.Fixed bug in applying mods stored as float values 2.Move Unit::m_spells to Creature::m_spells (used only for Pets and CreatureAI system) Note: Player have own m_spells field for diff. purpose. Update users. 3.Some cleanups
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1682 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Object.cpp
@@ -440,7 +440,7 @@ void Object::ApplyModUInt32Value(uint16 index, int32 val, bool apply)
 
 void Object::ApplyModFloatValue(uint16 index, float  val, bool apply)
 {
-    uint32 cur = GetFloatValue(index);
+    float cur = GetFloatValue(index);
     if(val > cur && !apply ) val = cur;
 
     SetFloatValue(index,cur+(apply ? val : -val));
